+++ Sources/HabitatCore/PolicyReasonCatalog.swift
@@ -127,6 +127,7 @@ enum PolicyReasonCatalog {
         .init(reasonCode: .dependencyResolutionMutation) { isSwiftPackageDependencyResolutionCommand($0) },
         .init(reasonCode: .remoteRepositoryAction) { isRemoteRepositoryActionCommand($0) },
         .init(reasonCode: .gitMutation) { isGitOrGitHubMutationGuard($0) },
+        .init(reasonCode: .userApprovalRequired) { isWorkspaceMutationCommand($0) },
         .init(reasonCode: .packageRegistryMutation) { isPackageRegistryMutationCommand($0) },
         .init(reasonCode: .packageManagerActivation) { isPackageManagerActivationCommand($0) },
         .init(reasonCode: .dependencyMutation) { isDependencyMutationCommand($0) },